Kill off hacky ctag
This wart doesn't exist in CHERI-RISC-V and we should not be introducing
it. Especially not by picking an encoding straight after the last (at
the time) one in the family, which of course now collides with the
newly-added csetequalexact. I don't like the non-standard
creadhpmcounter here either, that really should just be a normal
performance counter rather than a weird CHERI-specific thing, but it can
stay for now as it's not currently colliding (but woe betide it if it
does).